+++ changelog.md
@@ -7,6 +7,7 @@ This release works with version 0.27.0+ of UMC fw
 
 -   Support new Perpetual Tune feature
 -   Added Realtime Hashrate column to table
+-   Added User Target and Throttle Target in brackets in Perpetual Tune Target column
 
 ### Bug Fixes